general (working) OOD link
There's a "get an Open OnDemand terminal" link that at least for me, doesn't work on Chrome.  There's no other OOD mention in this section. 

I'm adding a link that I believe is the right one to the general OOD interface.  I'll track down if the terminal link is still useful or not.