Select an arbitrary port # for tunnel, use $RANDOM, not $UID
We need to give the user a port number for setting up tunneling. We tried putting in an arbitrary number, but multiple people just use that number literally. So I tried using $UID. Unfortunately, our UIDs are now too big for port numbers. There is a block of unreserved ports from 49152 to 65535, so instead of using UID, use some bash code to generate a random port number in the range 49152-59151.